Make CliGyroUI the default implementation of GyroUI
