+++ core/build.gradle
@@ -12,4 +12,5 @@ dependencies {
     implementation 'org.eclipse.aether:aether-impl:1.1.0'
     implementation 'org.eclipse.aether:aether-transport-file:1.1.0'
     implementation 'org.eclipse.aether:aether-transport-http:1.1.0'
+    implementation 'org.fusesource.jansi:jansi:1.16'
 }
